Add HandView and use shader file in FingerView
No good reason for doing these in a single commit, just bounced aroud a bit while doing them.  Handview displays a disc where the hand is.  Rendering is incomplete and not currently in use.  Purpose is to visualize finger-hand relationship in future interactions, and to orient fingers relative to hand to facilitate motion scaling without distording relative finger positions.