Use selected kit to obtain icons for DrumSelector and TrigView
This is done via the user settings object.  Also made changes in
application and main component init to ensure default values are
immediately written to user settings.  I *think* the settings object
does not require disk IO for reading values.  Otherwise this change is
horribly bad because it is done in the OpenGL rendering thread!  I'll
double check this soon but I needed the functionality ASAP for a preview
build.