Add a pattern subdirectory to be copied to app resources
I ran into an issue where the patterns folder did not exist inside my
app resources folder, so the pattern never got saved!  This folder
should be created when it is absent.  For now, I added a default folder
that can be copied to the app resources folder along with other
resources (this happens automatically during build on Mac)