Add pattern selector and wire up basic functionality.
Pattern selector is a WheelSelector that gets populated with the
contents of PatternManager.  When a pattern is saved-as, the selector
gets updated to reflect the new pattern list.

This commit also fixes a nasty crash on exit caused by the
MotionDispatcher getting destructed before the main component, which led
to a *new* instance of MotionDispatcher getting created later.  This had
prevented me from deleting any heap-allocated view objects, but once I
fixed this ordering problem I was able to call delete on all of them.
There are surely still some leaks on shutdown but this is a start.