Add "tracked finger missing" timer to all selection controls
HUD controls receive fingerEntered and fingerExited events, and use these
to track a single fingerView when fingerMotion events occur.  This is done
on a fist-in basis, so any other finger encountered in entered, exited, or
motion is ignored.  The problem is that fingerViews can suddenly
disappear, and when this happens fingerExited is never sent and the HUD
control is tracking a finger that doesn't exist.  As a result, the HUD
control stops responding.  The quick fix for this problem is to set a
timeout every time fingerMotion happens with the tracked finger.  If ever
the timeout completes and we haven't received motion events for the
tracked finger, we can assume that that finger is gone and we can reset
the tracking state.

It would be more optimal to fix the issue systemically, such that HUD
controls can be entirely ignorant to intermittent tracking.