Remove history logic from StrikeDetector
When this class was converted to a state machine, a history buffer was
added and used in the trigger detection logic.  This new step made
incorrect assumptions and broke the detection code in "Precision" and
"Balanced" Leap tracking mode, as well as in "High Speed" mode when the
sensitivity was set to anything below 0.5.  The purpose of the history
step was to reduce jitter.  Further work to that end should be done on a
branch and verified in all modes and multiple systems before committing to the main dev
branch.