Color instrument icons and pads
The the color pallete is saved with user settings and can be hand edited
by the artists.  Changed all green "on" state graphics to be a light
gray.  Color is rendered programatically.  Some changes to HUDButton and
View2d were required to allow for colorization of textures (before a
color of FFFFFFFF was used for textures, which is now the default and
can be changed).