Encapsulate texture info in struct TextureDescription, force power of 2 size
Forcing power of 2 square image size is easy enough--draw the image into
a new image with a power of 2 sized 'canvas' and load that new image as
the texture.  Then, the ratio of origional size to extended size is
required to properly map the texture onto surfaces.  To this I added a
struct to hold that info along with the texture id and use it
everywhere.

Extending the image size leads to huge memory wastage which could have
side effects.  I'll investigate ways to minimize the wasted memory.