Move some core initialization to the message thread
This was done when the splash screen was first added, and then removed
because it seemed unnecessary.  Later I realized that the cursor had stopped
working in release builds, and it turns out that delaying the
initialization fixes that.  Must be some kind of a race condition, and
it makes me nervous about other systems and platforms.  Need to dig a
bit deeper to find the root of the issue.