Use GLEW to determine VAO support
Parallels on my Mac Pro (which has a Radeon HD 4870), supposedly
supports OpenGL 2.1 but aparently does not support
GL_ARB_vertex_array_obect.  Because of this, GLBatch and GLTriangleBatch
function from GLTools crash my app as well as the Superbible 5 examples.
This change just logs whether VAOs are supported.  If not, the
"OPENGL_ES" macro can be defined in GLBatch.cpp and GLTriangleBatch.cpp,
which has the effect of disabling VAO functions.