Improve cursor tracking
The cursor previously selected an arbitrary finger, and now uses the
closest finger to the screen.  The selection only changes when the
tracked finger goes away (IOW if the tracked finger ceases to be the
closest one to the screen, it is still the selected finger until it
disappears).

Sometimes tracking data would stop intermittently, such as when the
pointing hand is close to the edge of the FOV.  This would cause the
cursor to flicker.  Now the cursor is dismissed after a timer, which
triggers .5 sec after the last frame of position data is encountered.

Also added a leaky integrator to generally smooth out the cursor motion.